Save as uint8 image

+++ highlight_finder.py
@@ -11,9 +11,9 @@ from skimage.measure import label, regionprops
 def main(fname):
     templates = dict(yellow=np.array([251, 240, 187]),
                      blue=np.array([210, 255, 245]))
-    im = imread(fname)
+    im = imread(fname).astype('int64')
     if im.shape[2] > 3:
-        im = rgba2rgb(im)*255
+        im = (rgba2rgb(im)*255).astype('int64')
     im_files = {}
     for key, template in templates.items():
         print('Testing {}'.format(key))
@@ -34,7 +34,7 @@ def save_highlight_extract(im, bounds, color, fname):
         tmp = im[minr:maxr, minc:maxc]
         f = '{}-{}-{}.png'.format(color, i, basename[:-4])
         fnames += [os.path.join(path, f)]
-        imsave(os.path.join(path, f), tmp)
+        imsave(os.path.join(path, f), tmp.astype('uint8'))
     return fnames
 
 def get_highlighted_regions(im, color, threshold=12):
